// src/domain/models/Patient.ts
// src/domain/models/Patient.ts
export type Patient = {
  id: string;
  identifier: string;
  active: string;
  name: string;
  gender: string;
  birthDate: string;
  deceased: string;
  address: string;
  phone: string;
  email: string;
  maritalStatus: string;
  multipleBirth: string;
  languages: string;
  contactName: string;
  contactPhone: string;
  managingOrg: string;
  generalPractitioner: string;
};